# 백엔드/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HUFS Town API
  version: 0.5.0
  description: GDG HUFS SSO, spaces, invitations, map editing and world admission. Media signaling uses the authenticated world WebSocket contract; Node control routes are private. Never log credentials, invite codes or admission headers.
paths:
...
  /api/v1/me/friends/requests/{requestId}/respond:
    parameters:
      - name: requestId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: respondToFriendRequest
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Only the target account can accept or decline a pending friend request. A decline applies a 24-hour retry cooldown for the original requester.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FriendRequestDecision" }
      responses:
        "200":
          description: Request decision.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [ACCEPTED, DECLINED] }
        "400": { $ref: "#/components/responses/AuthError" }
        "401": { $ref: "#/components/responses/AuthError" }
        "404": { $ref: "#/components/responses/AuthError" }
        "409": { $ref: "#/components/responses/AuthError" }
        "503": { $ref: "#/components/responses/AuthError" }
        default: { $ref: "#/components/responses/AuthError" }
...
  /api/v1/me/join-requests:
    get:
      operationId: listIncomingSocialJoinRequests
      security: [{ sessionCookie: [] }]
      description: Lists pending meetup requests and approved requests for seven days so the target can continue through ordinary space admission. A destination is included only because the requester explicitly selected and shared that space; the response never contains the target's current space, map, presence, or coordinates.
      responses:
        "200":
          description: Pending requests and recently approved destinations for the authenticated account.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SocialJoinRequest" }
        default: { $ref: "#/components/responses/AuthError" }
    post:
      operationId: createSocialJoinRequest
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Creates a seven-day asynchronous meetup request for the peer resolved server-side from a DIRECT conversation the caller belongs to. The caller must be a current member of the destination space and explicitly shares its name with the recipient. Group chats and conversations the caller does not belong to are rejected. Both directions of user blocking are enforced. Approval records consent only; the recipient must separately open the destination and pass its normal admission checks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conversationId, destinationSpaceId]
              properties:
                conversationId: { type: string, format: uuid }
                destinationSpaceId: { type: string, format: uuid }
                message: { type: string, maxLength: 160 }
      responses:
        "200":
          description: Existing pending request or newly created request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SocialJoinRequestState" }
        default: { $ref: "#/components/responses/AuthError" }
  /api/v1/me/join-requests/outgoing:
    get:
      operationId: listOutgoingSocialJoinRequests
      security: [{ sessionCookie: [] }]
      description: Lists the authenticated account's requests from the last seven days, including decisions and expiry.
      responses:
        "200":
          description: Recent requests sent by the authenticated account.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SocialJoinRequest" }
        default: { $ref: "#/components/responses/AuthError" }
  /api/v1/me/join-requests/{requestId}/respond:
    parameters:
      - name: requestId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: respondToSocialJoinRequest
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: Only the target account can approve or decline its pending request. Approval is a meetup consent signal and does not move either participant. The selected destination remains in the target's request list so they can separately open it through ordinary space admission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [APPROVE, DECLINE] }
      responses:
        "200":
          description: Request decision.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [APPROVED, DECLINED, EXPIRED] }
        default: { $ref: "#/components/responses/AuthError" }
  /api/v1/me/join-requests/{requestId}/cancel:
    parameters:
      - name: requestId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: cancelSocialJoinRequest
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: The requester may cancel their pending request.
      responses:
        "200":
          description: Request cancelled.
          content:
            application/json:
              schema:
                type: object
                required: [cancelled]
                properties:
                  cancelled: { type: boolean, const: true }
        default: { $ref: "#/components/responses/AuthError" }
...
components:
...
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: HUFS_TOWN_SESSION }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-TOKEN }
  schemas:
...
    FriendRequestDecision:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [ACCEPT, DECLINE] }
...
    SocialJoinRequest:
      type: object
      required:
        [
          id,
          displayName,
          message,
          status,
          requestedAt,
          expiresAt,
          destinationSpaceId,
          destinationSpaceName,
        ]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        message: { type: string, maxLength: 160 }
        status:
          {
            type: string,
            enum: [PENDING, APPROVED, DECLINED, CANCELLED, EXPIRED],
          }
        requestedAt: { type: integer, format: int64 }
        expiresAt: { type: integer, format: int64 }
        destinationSpaceId: { type: string, format: uuid, nullable: true }
        destinationSpaceName: { type: string, maxLength: 60, nullable: true }
    SocialJoinRequestState:
      type: object
      required: [id, status, requestedAt, expiresAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [PENDING] }
        requestedAt: { type: integer, format: int64 }
        expiresAt: { type: integer, format: int64 }
...
    AuthError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
  responses:
...
    AuthError:
      description: Authentication, policy, CSRF, rate limit, configuration or provider failure. No upstream body or secret is returned.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }